fix(importacao): corrige erros nas planilhas e adiciona nginx local
- importar_bens_consumo_xlsx: usa _cel() nas colunas 10/13/14 para evitar IndexError em linhas curtas
- contracts: muda cod_liquidacao de IntegerField para BigIntegerField (valores acima de 2.1bi na planilha)
- migration 0008: aplica alteração do BigIntegerField
- docker-compose: adiciona nginx na porta 80 como proxy reverso para acesso pela rede local
- nginx.local.conf: config do nginx para ambiente local (proxy + static/media)

+++ dimms/models/contracts.py
@@ -115,7 +115,7 @@ class Contrato(models.Model): #PRONTO
         verbose_name='N° SC',
     )
 
-    cod_liquidacao = models.IntegerField(
+    cod_liquidacao = models.BigIntegerField(
         blank=True,
         null=True,
         verbose_name='Código de Licitação',
